Add properties AxisCaptionX and AxisCaptionY
Define two new properties on the WpfGraphControl:
- AxisCaptionX
- AxisCaptionY
In the control template for WpfGraphControl, define a grid column for the Y caption and a row for the X caption and show the captions using TextBlock controls.
The new column and row auto size, and collapse to zero width or height respectively if the caption is empty.