Replace WpfGraphAxisControl.Ticks with .Divisions and add WpfGraphControl.DivisionsX and .DivisionsY
Replace the Ticks property on WpfGraphAxisControl with a Divisions property and define Ticks as Divisions+1. With this change, the Divisions property can have the same numeric value as the Rows and Columns properties on the WpfGraphGridLines control.

Add the properties DivisionsX and DivisionsY to WpfGraphControl.

Add TemplateBindings to the control template in Generic.xaml, to bind the DivisionsX and DivisionsY properties to the Rows and Columns of the WpfGraphGridLines control and to the Divisions properties of the respective WpfGraphAxisControl controls.