Add a margin for AutoY scaling.
Add a new abstract method ExpandRange to GraphDataPoint to expand the min and max values by a factor applied to the difference between the two values, i.e.
    min = min - factor * ( max - min )
    max = max + factor * ( max - min )

This will be used with the AutoY scaling option.

Obviously, it would be nice to make the factor configurable, but in this version it is hard coded as 0.05.

Note that the method ExpandRange works like a static method, in that it does not use the "this" reference at all.
It is not static, because it is abstract.